requirements.txt
streamlit
llama-cpp-python
PyMuPDF
